# src/main/java/com/egov/servicerequest/controller/ServiceApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.egov.servicerequest.controller;

import com.egov.servicerequest.model.ErrorRes;
import com.egov.servicerequest.model.ServiceDefinitionRequest;
import com.egov.servicerequest.model.ServiceDefinitionResponse;
import com.egov.servicerequest.model.ServiceDefinitionSearchRequest;
import com.egov.servicerequest.model.ServiceRequest;
import com.egov.servicerequest.model.ServiceResponse;
import com.egov.servicerequest.model.ServiceSearchRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-07-24T14:37:48.733516900+05:30[Asia/Calcutta]")
@Validated
@Tag(name = "Service Definition", description = "the Service Definition API")
@RequestMapping("${openapi.service-request.base-path:}")
public interface ServiceApi {

    default ServiceApiDelegate getDelegate() {
        return new ServiceApiDelegate() {};
    }

    /**
     * POST /service/definition/v1/_create
     *
     * @param serviceDefinitionRequest  (optional)
     * @return Request accepted. (status code 202)
     *         or Invalid input. (status code 400)
     */
    @Operation(
        operationId = "serviceDefinitionV1CreatePost",
        tags = { "Service Definition" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Request accepted.", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = ServiceDefinitionResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input.", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = ErrorRes.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/service/definition/v1/_create",
        produces = { "*/*" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ServiceDefinitionResponse> serviceDefinitionV1CreatePost(
        @Parameter(name = "ServiceDefinitionRequest", description = "") @Valid @RequestBody(required = false) ServiceDefinitionRequest serviceDefinitionRequest
    ) {
        return getDelegate().serviceDefinitionV1CreatePost(serviceDefinitionRequest);
    }


    /**
     * POST /service/definition/v1/_search
     *
     * @param serviceDefinitionSearchRequest  (optional)
     * @return Request accepted. (status code 202)
     *         or Invalid input. (status code 400)
     */
    @Operation(
        operationId = "serviceDefinitionV1SearchPost",
        tags = { "Service Definition" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Request accepted.", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = ServiceDefinitionResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input.", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = ErrorRes.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/service/definition/v1/_search",
        produces = { "*/*" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ServiceDefinitionResponse> serviceDefinitionV1SearchPost(
        @Parameter(name = "ServiceDefinitionSearchRequest", description = "") @Valid @RequestBody(required = false) ServiceDefinitionSearchRequest serviceDefinitionSearchRequest
    ) {
        return getDelegate().serviceDefinitionV1SearchPost(serviceDefinitionSearchRequest);
    }


    /**
     * POST /service/v1/_create
     *
     * @param serviceRequest  (optional)
     * @return Request accepted. (status code 202)
     *         or Invalid input. (status code 400)
     */
    @Operation(
        operationId = "serviceV1CreatePost",
        tags = { "Service" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Request accepted.", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = ServiceResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input.", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = ErrorRes.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/service/v1/_create",
        produces = { "*/*" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ServiceResponse> serviceV1CreatePost(
        @Parameter(name = "ServiceRequest", description = "") @Valid @RequestBody(required = false) ServiceRequest serviceRequest
    ) {
        return getDelegate().serviceV1CreatePost(serviceRequest);
    }


    /**
     * POST /service/v1/_search
     *
     * @param serviceSearchRequest  (optional)
     * @return Request accepted. (status code 202)
     *         or Invalid input. (status code 400)
     */
    @Operation(
        operationId = "serviceV1SearchPost",
        tags = { "Service" },
        responses = {
            @ApiResponse(responseCode = "202", description = "Request accepted.", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = ServiceResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input.", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = ErrorRes.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/service/v1/_search",
        produces = { "*/*" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ServiceResponse> serviceV1SearchPost(
        @Parameter(name = "ServiceSearchRequest", description = "") @Valid @RequestBody(required = false) ServiceSearchRequest serviceSearchRequest
    ) {
        return getDelegate().serviceV1SearchPost(serviceSearchRequest);
    }

}