# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: info@egovernments.org
    name: Digit
  description: "API specs for service attribute definition and capture service attribute\
    \ \ndetails \n"
  title: Service Request
  version: 1.0.0
servers:
- url: /
paths:
  /service/definition/v1/_create:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceDefinitionRequest'
      responses:
        "202":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ServiceDefinitionResponse'
          description: Request accepted.
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorRes'
          description: Invalid input.
      tags:
      - Service Definition
      x-content-type: application/json
      x-accepts: '*/*'
      x-tags:
      - tag: Service Definition
  /service/definition/v1/_search:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceDefinitionSearchRequest'
      responses:
        "202":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ServiceDefinitionResponse'
          description: Request accepted.
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorRes'
          description: Invalid input.
      tags:
      - Service Definition
      x-content-type: application/json
      x-accepts: '*/*'
      x-tags:
      - tag: Service Definition
  /service/v1/_create:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequest'
      responses:
        "202":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ServiceResponse'
          description: Request accepted.
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorRes'
          description: Invalid input.
      tags:
      - Service
      x-content-type: application/json
      x-accepts: '*/*'
      x-tags:
      - tag: Service
  /service/v1/_search:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceSearchRequest'
      responses:
        "202":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ServiceResponse'
          description: Request accepted.
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorRes'
          description: Invalid input.
      tags:
      - Service
      x-content-type: application/json
      x-accepts: '*/*'
      x-tags:
      - tag: Service
components:
  schemas:
...
    ServiceDefinitionSearchRequest:
      example:
        pagination:
          offSet: 5.637376656633329
          limit: 59.621339166831824
          sortBy: sortBy
          totalCount: 2.3021358869347655
          order: asc
        serviceDefinitionCriteria:
          code:
          - code
          - code
          clientId: clientId
          tenantId: pb.amritsar
          ids:
          - ids
          - ids
        requestInfo:
          userInfo:
            additionalroles:
            - roles:
              - name: name
                description: description
              - name: name
                description: description
              tenantId: tenantId
            - roles:
              - name: name
                description: description
              - name: name
                description: description
              tenantId: tenantId
            tenantId: tenantId
            mobile: mobile
            id: 6
            userName: userName
            email: email
            primaryrole:
            - name: name
              description: description
            - name: name
              description: description
          ver: ver
          requesterId: requesterId
          authToken: authToken
          action: action
          msgId: msgId
          correlationId: correlationId
          apiId: apiId
          did: did
          key: key
          ts: 0
      properties:
        requestInfo:
          $ref: '#/components/schemas/RequestInfo'
        serviceDefinitionCriteria:
          $ref: '#/components/schemas/ServiceDefinitionCriteria'
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
      - requestInfo
      - serviceDefinition
      type: object
    ServiceSearchRequest:
      example:
        pagination:
          offSet: 5.637376656633329
          limit: 59.621339166831824
          sortBy: sortBy
          totalCount: 2.3021358869347655
          order: asc
        serviceDefinition:
          accountId: accountId
          clientId: clientId
          referenceIds:
          - referenceIds
          - referenceIds
          tenantId: pb.amritsar
          ids:
          - ids
          - ids
          value: "{}"
          serviceDefIds:
          - serviceDefIds
          - serviceDefIds
        requestInfo:
          userInfo:
            additionalroles:
            - roles:
              - name: name
                description: description
              - name: name
                description: description
              tenantId: tenantId
            - roles:
              - name: name
                description: description
              - name: name
                description: description
              tenantId: tenantId
            tenantId: tenantId
            mobile: mobile
            id: 6
            userName: userName
            email: email
            primaryrole:
            - name: name
              description: description
            - name: name
              description: description
          ver: ver
          requesterId: requesterId
          authToken: authToken
          action: action
          msgId: msgId
          correlationId: correlationId
          apiId: apiId
          did: did
          key: key
          ts: 0
      properties:
        requestInfo:
          $ref: '#/components/schemas/RequestInfo'
        serviceDefinition:
          $ref: '#/components/schemas/ServiceCriteria'
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
      - requestInfo
      - serviceDefinition
      type: object
    ServiceDefinitionRequest:
      example:
        serviceDefinition:
          code: code
          clientId: clientId
          auditDetails:
            lastModifiedTime: 1
            createdBy: createdBy
            lastModifiedBy: lastModifiedBy
            createdTime: 6
          tenantId: pb.amritsar
          attributes:
          - regEx: regEx
            code: code
            dataType: String
            values:
            - values
            - values
            auditDetails:
              lastModifiedTime: 1
              createdBy: createdBy
              lastModifiedBy: lastModifiedBy
              createdTime: 6
            tenantId: pb.amritsar
            id: id
            isActive: true
            additionalDetails: "{}"
            required: true
            order: order
          - regEx: regEx
            code: code
            dataType: String
            values:
            - values
            - values
            auditDetails:
              lastModifiedTime: 1
              createdBy: createdBy
              lastModifiedBy: lastModifiedBy
              createdTime: 6
            tenantId: pb.amritsar
            id: id
            isActive: true
            additionalDetails: "{}"
            required: true
            order: order
          id: id
          isActive: true
          additionalDetails: "{}"
        requestInfo:
          userInfo:
            additionalroles:
            - roles:
              - name: name
                description: description
              - name: name
                description: description
              tenantId: tenantId
            - roles:
              - name: name
                description: description
              - name: name
                description: description
              tenantId: tenantId
            tenantId: tenantId
            mobile: mobile
            id: 6
            userName: userName
            email: email
            primaryrole:
            - name: name
              description: description
            - name: name
              description: description
          ver: ver
          requesterId: requesterId
          authToken: authToken
          action: action
          msgId: msgId
          correlationId: correlationId
          apiId: apiId
          did: did
          key: key
          ts: 0
      properties:
        requestInfo:
          $ref: '#/components/schemas/RequestInfo'
        serviceDefinition:
          $ref: '#/components/schemas/ServiceDefinition'
      required:
      - requestInfo
      - serviceDefinition
      type: object
    ServiceDefinitionResponse:
      example:
        pagination:
          offSet: 5.637376656633329
          limit: 59.621339166831824
          sortBy: sortBy
          totalCount: 2.3021358869347655
          order: asc
        serviceDefinition:
        - code: code
          clientId: clientId
          auditDetails:
            lastModifiedTime: 1
            createdBy: createdBy
            lastModifiedBy: lastModifiedBy
            createdTime: 6
          tenantId: pb.amritsar
          attributes:
          - regEx: regEx
            code: code
            dataType: String
            values:
            - values
            - values
            auditDetails:
              lastModifiedTime: 1
              createdBy: createdBy
              lastModifiedBy: lastModifiedBy
              createdTime: 6
            tenantId: pb.amritsar
            id: id
            isActive: true
            additionalDetails: "{}"
            required: true
            order: order
          - regEx: regEx
            code: code
            dataType: String
            values:
            - values
            - values
            auditDetails:
              lastModifiedTime: 1
              createdBy: createdBy
              lastModifiedBy: lastModifiedBy
              createdTime: 6
            tenantId: pb.amritsar
            id: id
            isActive: true
            additionalDetails: "{}"
            required: true
            order: order
          id: id
          isActive: true
          additionalDetails: "{}"
        - code: code
          clientId: clientId
          auditDetails:
            lastModifiedTime: 1
            createdBy: createdBy
            lastModifiedBy: lastModifiedBy
            createdTime: 6
          tenantId: pb.amritsar
          attributes:
          - regEx: regEx
            code: code
            dataType: String
            values:
            - values
            - values
            auditDetails:
              lastModifiedTime: 1
              createdBy: createdBy
              lastModifiedBy: lastModifiedBy
              createdTime: 6
            tenantId: pb.amritsar
            id: id
            isActive: true
            additionalDetails: "{}"
            required: true
            order: order
          - regEx: regEx
            code: code
            dataType: String
            values:
            - values
            - values
            auditDetails:
              lastModifiedTime: 1
              createdBy: createdBy
              lastModifiedBy: lastModifiedBy
              createdTime: 6
            tenantId: pb.amritsar
            id: id
            isActive: true
            additionalDetails: "{}"
            required: true
            order: order
          id: id
          isActive: true
          additionalDetails: "{}"
        responseInfo:
          ver: ver
          resMsgId: resMsgId
          msgId: msgId
          apiId: apiId
          ts: 0
          status: SUCCESSFUL
      properties:
        responseInfo:
          $ref: '#/components/schemas/ResponseInfo'
        serviceDefinition:
          items:
            $ref: '#/components/schemas/ServiceDefinition'
          type: array
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
    ServiceRequest:
      example:
        service:
          accountId: accountId
          clientId: clientId
          auditDetails:
            lastModifiedTime: 1
            createdBy: createdBy
            lastModifiedBy: lastModifiedBy
            createdTime: 6
          tenantId: pb.amritsar
          serviceDefId: serviceDefId
          attributes:
          - attributeCode: attributeCode
            auditDetails:
              lastModifiedTime: 1
              createdBy: createdBy
              lastModifiedBy: lastModifiedBy
              createdTime: 6
            id: id
            additionalDetails: "{}"
            value: "{}"
          - attributeCode: attributeCode
            auditDetails:
              lastModifiedTime: 1
              createdBy: createdBy
              lastModifiedBy: lastModifiedBy
              createdTime: 6
            id: id
            additionalDetails: "{}"
            value: "{}"
          id: id
          additionalDetails: "{}"
          referenceId: referenceId
        requestInfo:
          userInfo:
            additionalroles:
            - roles:
              - name: name
                description: description
              - name: name
                description: description
              tenantId: tenantId
            - roles:
              - name: name
                description: description
              - name: name
                description: description
              tenantId: tenantId
            tenantId: tenantId
            mobile: mobile
            id: 6
            userName: userName
            email: email
            primaryrole:
            - name: name
              description: description
            - name: name
              description: description
          ver: ver
          requesterId: requesterId
          authToken: authToken
          action: action
          msgId: msgId
          correlationId: correlationId
          apiId: apiId
          did: did
          key: key
          ts: 0
      properties:
        requestInfo:
          $ref: '#/components/schemas/RequestInfo'
        service:
          $ref: '#/components/schemas/Service'
      required:
      - requestInfo
      - serviceDefinition
      type: object
    ServiceResponse:
      example:
        pagination:
          offSet: 5.637376656633329
          limit: 59.621339166831824
          sortBy: sortBy
          totalCount: 2.3021358869347655
          order: asc
        service:
        - accountId: accountId
          clientId: clientId
          auditDetails:
            lastModifiedTime: 1
            createdBy: createdBy
            lastModifiedBy: lastModifiedBy
            createdTime: 6
          tenantId: pb.amritsar
          serviceDefId: serviceDefId
          attributes:
          - attributeCode: attributeCode
            auditDetails:
              lastModifiedTime: 1
              createdBy: createdBy
              lastModifiedBy: lastModifiedBy
              createdTime: 6
            id: id
            additionalDetails: "{}"
            value: "{}"
          - attributeCode: attributeCode
            auditDetails:
              lastModifiedTime: 1
              createdBy: createdBy
              lastModifiedBy: lastModifiedBy
              createdTime: 6
            id: id
            additionalDetails: "{}"
            value: "{}"
          id: id
          additionalDetails: "{}"
          referenceId: referenceId
        - accountId: accountId
          clientId: clientId
          auditDetails:
            lastModifiedTime: 1
            createdBy: createdBy
            lastModifiedBy: lastModifiedBy
            createdTime: 6
          tenantId: pb.amritsar
          serviceDefId: serviceDefId
          attributes:
          - attributeCode: attributeCode
            auditDetails:
              lastModifiedTime: 1
              createdBy: createdBy
              lastModifiedBy: lastModifiedBy
              createdTime: 6
            id: id
            additionalDetails: "{}"
            value: "{}"
          - attributeCode: attributeCode
            auditDetails:
              lastModifiedTime: 1
              createdBy: createdBy
              lastModifiedBy: lastModifiedBy
              createdTime: 6
            id: id
            additionalDetails: "{}"
            value: "{}"
          id: id
          additionalDetails: "{}"
          referenceId: referenceId
        responseInfo:
          ver: ver
          resMsgId: resMsgId
          msgId: msgId
          apiId: apiId
          ts: 0
          status: SUCCESSFUL
      properties:
        responseInfo:
          $ref: '#/components/schemas/ResponseInfo'
        service:
          items:
            $ref: '#/components/schemas/Service'
          type: array
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
    ErrorRes:
      description: "All APIs will return ErrorRes in case of failure which will carry\
        \ ResponseInfo as metadata and Error object as actual representation of error.\
        \ In case of bulk apis, some apis may chose to return the array of Error objects\
        \ to indicate individual failure."
      properties:
        ResponseInfo:
          $ref: '#/components/schemas/ResponseInfo'
        Errors:
          description: Error response array corresponding to Request Object array.
            In case of single object submission or _search related paths this may
            be an array of one error element
          items:
            $ref: '#/components/schemas/Error'
          minLength: 1
          type: array
      required:
      - ResponseInfo
      type: object
...
    RequestInfo:
      description: RequestInfo should be used to carry meta information about the
        requests to the server as described in the fields below. All eGov APIs will
        use requestinfo as a part of the request body to carry this meta information.
        Some of this information will be returned back from the server as part of
        the ResponseInfo in the response body to ensure correlation.
      example:
        userInfo:
          additionalroles:
          - roles:
            - name: name
              description: description
            - name: name
              description: description
            tenantId: tenantId
          - roles:
            - name: name
              description: description
            - name: name
              description: description
            tenantId: tenantId
          tenantId: tenantId
          mobile: mobile
          id: 6
          userName: userName
          email: email
          primaryrole:
          - name: name
            description: description
          - name: name
            description: description
        ver: ver
        requesterId: requesterId
        authToken: authToken
        action: action
        msgId: msgId
        correlationId: correlationId
        apiId: apiId
        did: did
        key: key
        ts: 0
      properties:
        apiId:
          description: unique API ID
          maxLength: 128
          type: string
        ver:
          description: API version - for HTTP based request this will be same as used
            in path
          maxLength: 32
          type: string
        ts:
          description: time in epoch
          format: int64
          type: integer
        action:
          description: "API action to be performed like _create, _update, _search\
            \ (denoting POST, PUT, GET) or _oauth etc"
          maxLength: 32
          type: string
        did:
          description: Device ID from which the API is called
          maxLength: 1024
          type: string
        key:
          description: API key (API key provided to the caller in case of server to
            server communication)
          maxLength: 256
          type: string
        msgId:
          description: Unique request message id from the caller
          maxLength: 256
          type: string
        requesterId:
          description: UserId of the user calling
          maxLength: 256
          type: string
        authToken:
          description: //session/jwt/saml token/oauth token - the usual value that
            would go into HTTP bearer token
          type: string
        userInfo:
          $ref: '#/components/schemas/UserInfo'
        correlationId:
          readOnly: true
          type: string
      required:
      - action
      - apiId
      - msgId
      - ts
      - ver
      type: object
...